requirements.txt
dvc 
dvc[gdrive]
sklearn~=0.0
pandas~=1.3.2
numpy~=1.21.2
pyyaml~=5.4.1
argparse~=1.4.0
scikit-learn~=0.24.2
joblib~=1.0.1
flask~=2.0.1
cassandra-driver~=3.25.0
pytest
tox
flake8